+++ includes/push-ui.php
@@ -60,7 +60,7 @@ function syndicatable() {
 		return false;
 	}
 
-	if ( ! in_array( get_post_type(), \Distributor\Utils\distributable_post_types(), true ) || ( ! empty( $_GET['post_type'] ) && 'dt_ext_connection' === $_GET['post_type'] ) ) {
+	if ( ! in_array( get_post_type(), \Distributor\Utils\distributable_post_types(), true ) || ( ! empty( $_GET['post_type'] ) && 'dt_ext_connection' === $_GET['post_type'] ) ) { // @codingStandardsIgnoreLine Nonce not required
 		return false;
 	}
 
@@ -303,6 +303,7 @@ function menu_content() {
 		<div id="distributor-push-wrapper">
 			<div class="inner">
 				<p class="syndicated-notice">
+					<?php /* translators: %s: post type name */ ?>
 					<?php printf( esc_html__( 'This %s has been distributed from', 'distributor' ), esc_html( strtolower( $post_type_object->labels->singular_name ) ) ); ?>
 					<a href="<?php echo esc_url( $site_url ); ?>"><?php echo esc_html( $blog_name ); ?></a>.
 
@@ -326,7 +327,7 @@ function menu_content() {
 		}
 
 		if ( ! empty( \Distributor\Connections::factory()->get_registered()['networkblog'] ) ) {
-			$sites = \Distributor\InternalConnections\NetworkSiteConnection::get_available_authorized_sites();
+			$sites = \Distributor\InternalConnections\NetworkSiteConnection::get_available_authorized_sites( 'push' );
 
 			foreach ( $sites as $site_array ) {
 				if ( in_array( $post->post_type, $site_array['post_types'], true ) ) {
@@ -357,7 +358,7 @@ function menu_content() {
 		$external_connections_query = new \WP_Query(
 			array(
 				'post_type'      => 'dt_ext_connection',
-				'posts_per_page' => 200,
+				'posts_per_page' => 200, // @codingStandardsIgnoreLine This high pagination limit is purposeful
 				'no_found_rows'  => true,
 				'post_status'    => 'publish',
 			)
@@ -365,7 +366,7 @@ function menu_content() {
 
 		$current_post_type = get_post_type();
 
-		if ( ! empty( $_GET['post_type'] ) ) {
+		if ( ! empty( $_GET['post_type'] ) ) { // @codingStandardsIgnoreLine nonce not required
 			$current_post_type = sanitize_key( $_GET['post_type'] );
 		}
 
@@ -443,6 +444,7 @@ function menu_content() {
 			<div class="inner">
 
 				<?php if ( ! empty( $dom_connections ) ) : ?>
+					<?php /* translators: %s the post title */ ?>
 					<p><?php echo sprintf( esc_html__( 'Distribute &quot;%s&quot; to other connections.', 'distributor' ), esc_html( get_the_title( $post->ID ) ) ); ?></p>
 
 					<div class="connections-selector">