+++ tests/wpassure/bootstrap.php
@@ -1,8 +1,0 @@
-<?php
-/**
- * Bootstrap WP Assure tests
- *
- * @package  distributor
- */
-
-require_once( __DIR__ . '/includes/TestCase.php' );